requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
dj-database-url>=2.1
gunicorn>=21.2
whitenoise>=6.6
psycopg2-binary>=2.9
